adds requirements.txt (resolves #4) - removes google dependencies for now - fixes compatibility for newer versions of PyQt5

+++ widgets/calendar/all_day_widget.py
@@ -115,7 +115,7 @@ class AllDayWidget(TimelineWidget):
         painter = QPainter(self)
         for d in range(0, self.days):
             painter.setFont(QFont('Calibri',
-                                  min(self.bg_widgets[d].width() / 8, 16)
+                                  min(int(self.bg_widgets[d].width() / 8), 16)
                                   )
                             )
             day = (self.start_date + timedelta(days=d))
